Move NN files to a subfolder.
