+++ src/Makefile.am
@@ -86,7 +86,7 @@ GRIDCOIN_CORE_H = \
 	main.h \
 	miner.h \
 	mruset.h \
-	neuralnet.h \
+	neuralnet/neuralnet.h \
 	netbase.h \
 	net.h \
 	pbkdf2.h \
@@ -177,7 +177,7 @@ libgridcoin_util_a-version.$(OBJEXT): obj/build.h
 libgridcoin_util_a_CPPFLAGS = $(AM_CPPFLAGS) $(GRIDCOIN_INCLUDES)
 libgridcoin_util_a_CXXFLAGS = $(AM_CXXFLAGS) $(PIE_FLAGS)
 libgridcoin_util_a_SOURCES = $(GRIDCOIN_CORE_CPP) \
-    neuralnet_stub.cpp \
+    neuralnet/neuralnet_stub.cpp \
     $(GRIDCOIN_CORE_H)
 
 if TARGET_WINDOWS
