+++ src/main.cpp
@@ -21,7 +21,7 @@
 #include "boinc.h"
 #include "beacon.h"
 #include "miner.h"
-#include "neuralnet.h"
+#include "neuralnet/neuralnet.h"
 #include "backup.h"
 #include "appcache.h"
 #include "tally.h"
