+++ src/neuralnet/neuralnet_stub.cpp
